+++ test/login.ts
@@ -1,9 +1,19 @@
 import { ADTClient } from "../src"
+import { AdtHTTP } from "../src/AdtHTTP"
 
 export function create() {
   return new ADTClient(
     process.env.ADT_URL!,
     process.env.ADT_USER!,
     process.env.ADT_PASS!
+  )
+}
+export function createHttp(language: string = "") {
+  return new AdtHTTP(
+    process.env.ADT_URL!,
+    process.env.ADT_USER!,
+    process.env.ADT_PASS!,
+    "",
+    language
   )
 }